requirements.txt
pickle
numpy-1.26.4
pydub-0.25.1
pytest-8.2.2